app/requirements.txt
redis
yfinance
boto3
psycopg2-binary